% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battleship.R
\name{play}
\alias{play}
\title{Play the battleship game}
\usage{
play(rows, columns)
}
\arguments{
\item{rows}{number of rows for your board}

\item{columns}{number of columns for your board}
}
\value{
\link{X} you missed and no parts of the ship were struck

\link{O} you struck part of the ship
}
\description{
Randomly hides a ship of size 2 or 3 onto a board size of m rows and n columns and it's down to you to find the ship!
}
\examples{
play(5,5) creates a 5x5 board
}
\author{
Peter Chang
}
