% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battleship.R
\name{play}
\alias{play}
\title{Play the battleship game}
\usage{
play(rows, columns)
}
\arguments{
\item{rows}{number of rows for your board}

\item{columns}{number of columns for your board}
}
\description{
Randomly hides a ship onto a board size of m rows and n columns and it's down to you to find the ship!
}
\author{
Peter Chang
}
